requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
